Set up halfway points for levers
For the levers, set up halfway angles for them to reset to once the player has let go of them - so no as to return a constant value.
Also set up a spring for the wheel and the levers, which will automatically reset the joints to the angles mentioned above.